+++ jipp-core/src/main/java/com/hp/jipp/cups/Cups.kt
@@ -1,0 +1,4 @@
+package com.hp.jipp.cups
+
+object Cups {
+}