+++ HISTORY.md
@@ -1,5 +1,8 @@
 # History of Changes
 
+## 0.7.7
+* Allow stream reads/writes of AttributeGroup (#103).
+
 ## 0.7.6
 * Fetch CUPS attributes from https://www.cups.org/doc/spec-ipp.html (#95).
 * Handle copy count and blank page insertion in `PclmWriter` and `PwgWriter` (#99).